move resource view to content app
